reorganize util functions
